+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    order: mark test to run in a specific order
+
