+++ scripts/test_install.sh
@@ -1,0 +1,10 @@
+#!/bin/zsh
+
+export CI=true
+
+# Run the main install script
+./scripts/install.sh
+
+# Additional setup for testing (if any)
+# e.g., creating test directories, setting up mock data, etc.
+
