+++ scripts/test_uninstall.sh
@@ -1,0 +1,10 @@
+#!/bin/zsh
+
+export CI=true
+
+# Run the main uninstall script
+./scripts/uninstall.sh
+
+# Additional cleanup for testing (if any)
+# e.g., removing test directories, deleting mock data, etc.
+
